Sacred texts in tidy formats

Sacred texts in tidy formats

Datasets

king_james_version tibble of the King James bible, one row per verse.

apocrypha tibble of books in the Apocrypha, one row per verse.

vulgate tibble of books in the Vulgate, one row per verse.

tanach tibble of books in the Tanach, one row per verse.

septuagint tibble of books in the Septuagint, one row per verse.

greek_new_testament tibble of books in the Greek New Testatement, one row per verse.

middle_english_stopwords tibble of Middle English stopwords.

middle_english_sentiments tibble lexicon of Middle English positive and negative terms.